pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agit"
version = "0.1.0"
description = "Git-like version control for AI agents"
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
dependencies = [
    "agit-core",
    "typer>=0.9",
    "rich>=13.0",
]

[project.optional-dependencies]
google-adk = ["google-adk>=1.25"]
openai = ["openai-agents>=0.6"]
claude = ["claude-agent-sdk>=0.1.41"]
vercel = ["ai>=6.0"]
langgraph = ["langgraph>=1.0"]
crewai = ["crewai>=1.9"]
mcp = ["mcp>=2.1", "fastmcp>=2.1"]
ui = ["streamlit>=1.30", "plotly>=5.18"]
observability = ["prometheus-client>=0.20", "opentelemetry-api>=1.20"]
postgres = ["asyncpg>=0.29"]
s3 = ["boto3>=1.34"]
dev = ["pytest>=8.0", "pytest-cov", "pytest-asyncio", "mypy>=1.8", "ruff>=0.2"]
all = ["agit[google-adk,openai,claude,vercel,langgraph,crewai,mcp,ui,observability,postgres,s3,dev]"]

[project.scripts]
agit = "agit.cli.app:main"

[tool.hatch.build.targets.wheel]
packages = ["python/agit"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
strict = true
